TST/REF: misplaced frame.indexing tests (#37319)
