+++ pandas/tests/frame/indexing/test_setitem.py
@@ -8,10 +8,12 @@ from pandas import (
     DataFrame,
     Index,
     Interval,
+    NaT,
     Period,
     Series,
     Timestamp,
     date_range,
+    notna,
 )
 import pandas._testing as tm
 from pandas.core.arrays import SparseArray
@@ -180,3 +182,34 @@ class TestDataFrameSetItem:
         df["obj"] = obj
 
         tm.assert_frame_equal(df, expected)
+
+    def test_setitem_dt64tz(self, timezone_frame):
+
+        df = timezone_frame
+        idx = df["B"].rename("foo")
+
+        # setitem
+        df["C"] = idx
+        tm.assert_series_equal(df["C"], Series(idx, name="C"))
+
+        df["D"] = "foo"
+        df["D"] = idx
+        tm.assert_series_equal(df["D"], Series(idx, name="D"))
+        del df["D"]
+
+        # assert that A & C are not sharing the same base (e.g. they
+        # are copies)
+        b1 = df._mgr.blocks[1]
+        b2 = df._mgr.blocks[2]
+        tm.assert_extension_array_equal(b1.values, b2.values)
+        b1base = b1.values._data.base
+        b2base = b2.values._data.base
+        assert b1base is None or (id(b1base) != id(b2base))
+
+        # with nan
+        df2 = df.copy()
+        df2.iloc[1, 1] = NaT
+        df2.iloc[1, 2] = NaT
+        result = df2["B"]
+        tm.assert_series_equal(notna(result), Series([True, False, True], name="B"))
+        tm.assert_series_equal(df2.dtypes, df.dtypes)
